fixed minor runtime bugs

+++ src/args.py
@@ -4,7 +4,7 @@ from dataclasses import dataclass
 class RuntimeArgs:
     llama_endpoint: str
     models: list[str] | None
-    fps: int
+    fps: float
     allow_single_frame: bool
     model: str
     temperature: float
